+++ crates/compiler/builtins/bitcode/Cargo.toml
@@ -11,6 +11,7 @@ version.workspace = true
 tempfile.workspace = true
 
 [build-dependencies]
+roc_command_utils = { path = "../../../utils/command" }
 # dunce can be removed once ziglang/zig#5109 is fixed
 dunce = "1.0.3"
