+++ crates/wasi-libc-sys/build.rs
@@ -1,4 +1,4 @@
-use roc_utils::zig;
+use roc_command_utils::zig;
 use std::env;
 use std::ffi::OsString;
 use std::fs;
